rename: profile icon 경로 변경
