--- src/components/icon/Son_profile.jsx
@@ -1,51 +1,0 @@
-import React from 'react';
-import {Circle, Ellipse, G, Mask, Path, Svg} from 'react-native-svg';
-
-export const Son_profile = () => {
-  return (
-    <Svg
-      width="39"
-      height="39"
-      viewBox="0 0 39 39"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg">
-      <Circle cx="19.6699" cy="19.5" r="19" fill="#4D83F4" />
-      <Mask
-        id="mask0_102_2550"
-        style="mask-type:alpha"
-        maskUnits="userSpaceOnUse"
-        x="0"
-        y="0"
-        width="39"
-        height="39">
-        <Circle cx="19.6699" cy="19.5" r="19" fill="#1D1D45" />
-      </Mask>
-      <G mask="url(#mask0_102_2550)">
-        <Path
-          d="M19.732 11.0157C29.5828 11.6754 30.3435 20.618 28.771 24.8468C27.2992 28.8048 22.7542 29.8009 20.5217 29.7205C15.9715 29.5565 12.83 26.7868 12.1829 25.32C11.6383 24.0855 10.9343 21.8951 10.6892 19.9797C10.5168 18.6325 10.0152 17.285 9.15533 16.2337C6.27504 12.7123 6.57406 9.19996 7.11491 7.92907C9.23602 2.9449 11.7479 4.58615 12.5266 6.30829C14.3388 10.3162 18.2373 10.9156 19.732 11.0157Z"
-          fill="black"
-        />
-        <Path
-          d="M20.341 33.3822C13.3974 32.8096 12.1431 24.7872 12.392 20.0984C12.4192 19.5865 12.8511 19.1992 13.3636 19.1992H27.2464C27.769 19.1992 28.2055 19.6052 28.2317 20.1271C28.787 31.177 23.4707 33.5992 20.341 33.3822Z"
-          fill="#FFCEC3"
-        />
-        <Ellipse
-          cx="12.3243"
-          cy="24.8995"
-          rx="1.71915"
-          ry="2.38771"
-          transform="rotate(-21.0616 12.3243 24.8995)"
-          fill="#FFCEC3"
-        />
-        <Ellipse
-          cx="1.71915"
-          cy="2.38771"
-          rx="1.71915"
-          ry="2.38771"
-          transform="matrix(-0.968757 -0.248014 -0.248014 0.968757 30.8253 23.0127)"
-          fill="#FFCEC3"
-        />
-      </G>
-    </Svg>
-  );
-};
